iBurnService : preserve internal app data (ignore last commit)
iBurnService.UpgradeLifeboat is an interface for saving and restoring
internal app data before and after an iBurn API data refresh.

Add SimpleLifeboat to simply favorites state based on playaId. This is
suitable or Art and Camps.

Add EventLifeboat to persist favorite state based on playaId and start
date, for events that have multiple occurrences with the same playaId.

This seems a good alternative to generating a deterministic UUID for
all item types based on the PlayaEvents data scheme (another id, yuck)